[refactor] 내체험관리 컴포넌트로 파일이동
