+++ app/mypage/page.tsx
@@ -7,15 +7,25 @@ import ReservationList from '../components/reservation-list';
 import Navbar from '@/components/common/navbar';
 import ReservationCalendar from '../components/reservation-calendar';
 import Footer from '@/components/common/footer';
+import MyActivities from '../../components/myactivities/myactivities';
+import ActivitiesRegister from '../../components/myactivities/activities-register';
 
 export default function Page() {
   const [selectedMenu, setSelectedMenu] = useState('myinfo');
   const [isMobile, setIsMobile] = useState<boolean | null>(null);
   const [isOpen, setIsOpen] = useState(false);
+  const [key, setKey] = useState(0);
 
   function getPageSize(width: number): boolean {
     return width < 745;
   }
+
+  const handleSelectMenu = (menuId: string) => {
+    if (menuId === 'treatReservation' && menuId === selectedMenu) {
+      setKey(prevKey => prevKey + 1);
+    }
+    setSelectedMenu(menuId);
+  };
 
   useEffect(() => {
     const initialIsMobil = getPageSize(document.documentElement.clientWidth);
@@ -51,10 +61,9 @@ export default function Page() {
       case 'reserveList':
         return <ReservationList />;
       case 'treatReservation':
-        return <div>내 체험 관리 컴포넌트</div>;
-        {
-          /*여기 컴포넌트 갈아끼워야해요*/
-        }
+        return <MyActivities />;
+      case 'activitiesRegister':
+        return <ActivitiesRegister />;
       case 'reserveCalendar':
         return <ReservationCalendar />;
         {
@@ -77,12 +86,14 @@ export default function Page() {
       {isMobile ? (
         // **모바일 환경**
         <div>
-          <SideNavi selectedMenu={selectedMenu} onSelectMenu={setSelectedMenu} isMobile={isMobile} onOpenModal={() => setIsOpen(true)} />
+          <SideNavi selectedMenu={selectedMenu} onSelectMenu={handleSelectMenu} isMobile={isMobile} onOpenModal={() => setIsOpen(true)} />
           {isOpen && (
             <OverlayContainer>
               <div className="h-full w-full overflow-y-auto bg-white">
                 <Navbar />
-                <div className="px-4 pt-6">{renderContent()}</div>
+                <div className="px-4 pt-6" key={key}>
+                  {renderContent()}
+                </div>
                 <Footer />
               </div>
             </OverlayContainer>
@@ -91,8 +102,10 @@ export default function Page() {
       ) : (
         // **PC/태블릿 환경**
         <div className="flex overflow-y-auto tablet:gap-4 pc:gap-6">
-          <SideNavi selectedMenu={selectedMenu} onSelectMenu={setSelectedMenu} />
-          <div className="flex-grow">{renderContent()}</div>
+          <SideNavi selectedMenu={selectedMenu} onSelectMenu={handleSelectMenu} />
+          <div className="flex-grow" key={key}>
+            {renderContent()}
+          </div>
         </div>
       )}
     </div>
